refactor: move CopyInterner tests into copy.rs
